made an effect when hovering over the image on my webpage
check out my webpage–hovering the mouse over it causes a blue light to form along the image. i can work to make it an animation so that the light fades in and out instead of just appearing and disappearing. we can also use this on the main page after we get all the photos of us. we can hover over each photo and produce the same effect, so that users will know to click the photos in order to access each of our individual webpages.